Makefile: build static binaries for chargeback and promsum
Before this change the chargeback and promsum binaries needed to be
compiled on a system with musl libc, or they would fail to find their
libc in the alpine based container image. This commit changes the
Makefile to statically compile chargeback and promsum, thus allowing
them to be built on glibc based systems.