pkg/chargeback: Set Report Status to error if Report is started
If we find an already started report, it can mean only a few things:
- Report generation failed and the operator was unable to set its status
to error
- Report generation succeeded, and the operator was unable to set it's
status to finished
- There is more then 1 chargeback operator running at the same time.

We largely assume 1 chargeback operator is running, but don't enforce
this, but will in the future by using leader election, so we can also
ignore that case.  So until we can introspect the state, we have to
assume the report generation failed.