hack,Makefile: Run codegen on builds
This adds 2 scripts for helping in determining the targets and
dependencies of the k8s-update-codegen Makefile target. This allows
adding the codegen target as a dependency to our build targets, without
forcing a rebuild everytime. This is done using regular Make behavior
by producing a list of dependency files and target files for Make to use
when checking modification timestamps to determine if a recipe needs to
be ran.