pkg/chargeback: Add and use hiveQueryer interface which handles disconnects
Re-using connections is great accept the thrift client does not handle
disconnects, so this updates all code to use a hiveQueryer which only
knows how to query hive. The underlying implementation caches the thrift
client and initializes it when needed. If specific errors indicating the
connection has been lost occur when querying, the hiveQueryer
implementation will try to create a new connection and re-run the query.

This is particularly important if hive restarts while running.

Also updates chargeback operator to configure the query logging based on
an env var.