pkg/hive: Support adding new partitions to AWS Billing datastore table
Since we cannot introspect the table yet, and we haven't decided on
storing the partitions we've created in a CRD/ConfigMap, this will
handle updating the table partitions by dropping existing ones and
adding partitions based on the billing manifests at that time. In the
future, we'll compute a set of partitions to remove and a set to add
instead of deleting them all.

Partitions are on the billing_period_start and billing_period_end
columns, ensuring we only have one partition for a given billing period
at a time.

The partition columns are string types, which is what's best-practice
for hive apparently. This is necessary also to be able to delete the
partitions, as there is buggy/odd behavior in Hive when deleting the
default partition when the partition keys aren't all strings.