pkg/chargeback: Requeue already started reports more carefully
Compare the UUID of reports from cache to the UUID in the API when
deciding if we should requeue.

Also pass the correct report from the API object when updating the
cache. Previously we were re queueing the same object which resulted in
the same reporting getting processed over and over due to us overwriting
the report in cache preventing the status from ever becoming
not-started.