pkg/{hive,aws}: Assume single directory for each month of billing data
Based on our existing billing reports in AWS and the docs, we shouldn't
expect to have more than one directory for a given month of billing data
ever. There are multiple assembly directories containing billing data
for a given billing period, but we should only ever be using one at a
time.