pkg/{hive,chargeback}: Pass time objects to generation queries, and use funcs to format to timestamp
Updates manifests to use use new funcs. This is needed to support
the aws billing report which uses strings for partition columns and they
have a different timestamp style format being used.

Rather than passing in more variables, we can pass the time object
directly and use template functions to format based on how the time is
being used in the query.