prom-queries: Fix collection of metrics
- Sums categorically by pod, namespace, and node, which gets the total usage for the
whole pod, summing metrics for all containers in each pod. This fixes
another duplicate timestamp issue, as multiple containers in a pod
resulted in multiple timestamps existing for the same pod.
- No longer divides by the node capacity for the metric. This no longer
makes sense as it gives us a ratio of usage for a specific node, and our
presto queries expect to be aggregating bytes or cores, not percentages.