manifests: Update manifests to conform to new API schema
- Promsum Datastores only take a single Promtheus query
- ReportGenerationQueries have been updated to use a list of Datastores.
- ReportGenerationQueries no configure views. Existing report queries
still have views disabled as they're not re-usable by other queries in
the current form.
- Datastore names have been updated to be more specific about the metric
they're storing, and memory limits and cpu core limits have been added
as Datastores

As part of this, unnecessary parts of queries have been removed. To
compute the total usage over a period of time and get the rate does not
require window functions, and can be done with a simple sum of the
metric multiplied by the time.