pkg: don't run reports until the end time has been reached
This commit changes chargeback to by default wait to run reports until
the end time + a configurable grace period has been reached. This allows
for reports to be scheduled to run in the future.

This behavior can be disabled for a given report with the
`runImmediately` field, and the grace period can be modified for a given
report with the `gracePeriod` field. The default grace period is 5
minutes.