Makefile/Jenkins: Handle docker image build/tag/push for releases better
This refactoring eliminates some copy/paste boiler-plate in the
Makefile, but also makes it possible to just tag images and push those
tags. During releases we would like to avoid re-building images, and
instead just tag existing ones. The Jenkinsfile now checks for a boolean
parameter to determine if it should use this logic, and then attempts to
find a matching git tag for the commit to use when tagging and pushing
images.