Makefile: Generate make targets and use them instead of looping
The *-all targets previously used a foreach loop to remove some
copy/paste, but this meant make couldn't parallelize the contents of the
loop with jobs. To fix this, instead we generate a make target
and then use the generated list of targets as a dependency of the *-all
target, allowing make to natively see each as a separate target it can
run in a job.