charts: Fix storagelocation rendering
Helm was merging the overriden defaultStorage spec with the default,
resulting in both local and s3 being specified, resulting in only local
being used.

Add a `type` key to allow the template to differentiate between what's
being used and template each differently based on the type, to avoid
the merging issues with yaml.