charts/chargeback-operator: Handle multiple kube_node_info metrics for same node
If a node updates then the labels on the metric change, and this results
in there being a many to many relationship, which breaks the group_left
vector matching operator.

Fix it by using the max() aggregation function on the node and
provider_id labels to ensure we only ever get one vector for a given
node.