manifets/alm: Update deployment in chargeback CSV for latest ALM
Icon is no longer needed, as it's not required anymore, and additionally
chargeback is not going to be displayed in the regular OCS catalog, so
the icon serves no purpose.

ALM now takes a full deployment object in the deployments section, so
this mostly updates it to include a name/spec for the deployment.