charts/chargeback-operator: Fix pod cpu/memory usage PrometheusQueries
The pod_ip, node, and host_ip labels can be empty from the kube_pod_info
metric, meaning that matching on the pod, namespace labels isn't
necessarily enough to match to a single vector from kube_pod_info.

This happens when a pod isn't scheduled, or fully initialized yet, and
kube-state-metrics sees the pod state with those fields unset.

To handle this, we exclude the kube_pod_info time series where those
labels are empty, since we're only using it to add the node name into
the final result.