Remove release scripts, they're not really needed
Back when we were in a pre-alpha state, these release scripts where
helpful in getting people the docs and scripts needed to install and run
Chargeback. After we started publishing docs to coreos.com this wasn't
as necessarily, and now that we're about to be open source, they're
completely unneeded.