pkg/chargeback,charts: Replace hiveAWSPartitionTimestamp template function with billingPeriodTimestamp
The timestamp used here is actually not hive specific, and its actually
just how the AWS billing period is represented in both the bucket and in
the timestamps within the JSON manifest. So instead of using a confusing
name, remove it, and replace it's usage with something more descriptive
to what it's actually for.