hack: Install ClusterRole/ClusterRoleBinding granting access to Prometheus in Openshift
Cluster monitoring has an RBAC based auth proxy which requires our
service-account to have access to GET namespaces. To handle this in the
short-term this is handled in hack/openshift-install.sh and
hack/openshift-uninstall.sh.

For context, the follow link is where the source of the RBAC permissions
required is located:
https://github.com/openshift/cluster-monitoring-operator/blob/52105c47f7dc4497664a77668269a3d499f85b6b/assets/prometheus-k8s/prometheus-k8s.yaml#L60