pkg/chargeback: Read from presto for readiness check after initialized
The original intent of the readiness endpoint was to check if we're
initialized and can query Presto. Constantly inserting for due to
frequent readiness probes sometimes causes the Presto query queue to get
blocked up, and INSERTs cost more than SELECTs.

Now that we have a health check also running INSERTSs, we can stop
INSERTing for readiness after the initialization.

This PR updates the readiness and health check endpoints to do the
following:

- The operator is not ready until it's initialized, meaning it's caches
and API watches are working, and it's inserted into Presto at least
once.
- After it's initialized, readiness does a SELECT query to verify it can
still talk to Presto
- Health checks always perform insert queries.