*: Separate out Jenkinsfile logic into multiple jobs
This breaks up the Jenkinsfile into multiple Jenkinsfile's in the
jenkins/ directory to allow re-use of logic for jobs, and to allow
re-running individual portions more easily.

A new addition, the jenkins/deploy.groovy job is designed to allow
typical deploy logic without the same amount of overrides needed to run
e2e and integration.