pkg/{apis,chargeback,hive} Replace S3 storage with more generic hive storage config
This updates the Storage CR to be more agnostic so that it accepts most
parameters and properties to creating a table allowing significantly
more flexibility for storage. The default values are still setup to HDFS
by default, so the overall behavior is unchanged, except that existing
configurations need to be updated to the new style.