pkg/chargeback: Refactor http.go to rely on *Chargeback less
Beginnings of refactoring that reduce the overall interface required for
the HTTP API to work. This will be necessary in the tests that we are
adding for the HTTP API so that can control the behavior as needed.

Instead of using the Chargeback object, we instead pass each dependency
to the newRouter constructor to allow us to stub or mock the
implementations being used more easily.