pkg/{presto,chargeback}: Refactor presto querying in Chargeback to use presto ExecQueryer interface
Updates querying logic to rely on presto.ExecQueryer instead of
db.Queryer. The presto.ExecQueryer is a higher level abstraction over
the db.Queryer interface which has an implementation in presto.(*DB).
This higher level abstraction will be easier to fake/mock in tests since
we can avoid having to make the sql.Rows and instead just mock creating
presto.Row objects which are easier to construct.