charts/metering-operator: Support configuring openshift auth proxy to auth metering service
Adds a set of configuration options for adding an extra sidecar
container to the metering pod which is responsible for authenticating
incoming requests to the metering pod and proxying them to the metering
operator container. The auth proxy uses openshift service accounts for
auth and requires that the service account has namespace access to be
allowed to access the metering API. To enable verification of
permissions metering-operator.authProxy.subjectAccessReviewEnabled must
be enabled and the metering service account must have permissions to
create SubjectAccessReview resources (requires a ClusterRole and
ClusterRoleBinding). Currently the metering-helm-operator has support
for creating these RBAC resources, but requires these permissons itself,
which is not yet supported.