pkg/operator: Handle deletes using finalizers and drop tables on deletes
For Reports, ScheduledReports, and ReportDataSources, this updates
the ScheduledReport and ReportDataSource workers to use finalizers to
handle cleanup of any running Go routines processing their tasks, and
then for all of the, we use Finalizers on PrestoTable resources and an
updated PrestoTable handler to process deletes on PrestoTables by
dropping the associated tables.

As part of the updates to the PrestoTable handler, we switch to doing
partition updates in the ReportDataSource worker and use the resync
interval for updating PrestoTable periodically.