Documentation: roll-up report guide & explanation
Using already built-in macros and a custom report generation query, it
is possible to create roll-up reports — reports that are built out of
other reports. An initial guide for users to set up a test report and a
brief explanation of how they work provides a potential base for future
included roll-up reports.