charts,images: Remove separate metering-presto image
Move all configuration files from Presto image into Presto chart
avoiding the need for the metering-presto image

This also required
https://github.com/operator-framework/presto/commit/ae85f2d7cd4e2fac275d64664e8e5a29a9e7477a
to the Presto base image. That commit which sets the user to a non-root
UID (1003) by default, and ensures /etc/passwd is writable by the
default user, or a process with an EGID of 0. This is required because
we set runAsNonRoot in the Presto deployment. /etc/passwd must be
writable to GID 0 for the entrypoint to add the current user's to
/etc/passwd when the UID changes (like in Openshift).