pkg/operator: Refactor query validation and remove redundancies
This code was written a bit hastily last time and when I wrote this
there was a lot of duplication and weird logic due to it being used
slightly differently in different places. Usage is converged on
GetAndValidateGenerationQueryDependencies which means some cleanup can
be done. This will also make testing easier, as there is less overall
functions to test.