pkg/operator: Fix ReportGenerator interface and use it for report execution
After the initial implementation was finished the interface wasn't in
use yet, and didn't correctly match the implementation. Now it's in use
by the Reporting struct and additionally, the deleteExistingData
parameter was added back into the GenerateReport method signature to
allow using this single implementation for Reports and ScheduledReports
directly.