pkg/operator: Remove Informer types from Reporting type
Remove all of the metering type specific informer factories that were
unused, and also ReportInformer. ReportInformer is not used but is being
removed because we shouldn't need to update our internal indexer store
when we notice a Report has changed since we fetched it. Upstream
controllers also never directly modify the internal indexer store so
this seems like bad-practice in addition to it not being necessary.

Instead we can rely on requeues to eventually pick up on the new
resources values.