charts: Set JVM options for hdfs, hive and presto
Use ExitOnOutOfMemoryError instead of OnOutOfMemoryError=kill -9 %p
which doesn't work in a container when the JVM is PID 1 which can't be
killed. Also updates hdfs/hive to use the G1 garbage collector and be
verbose when there are GC pauses or errors. Also enables JMX options.
Taken from hadoop-env.sh and hive-env.sh.