manifests,pkg,cmd: Improve performance when importing new Prometheus ReportDataSource data
When first handling a ReportDataSource it's likely there is already a
good amount of data available, so we should attempt to collect it. To
handle this, we need to be able to query for more data at once, and
store more at once as well. To handle this reasonably, most of this is
now exposed as configuration, to allow improving performance if more
resources are available.

This also adds status fields to record the most recent import, the
oldest metric and the newest metric imported for each Prometheus
ReportDataSource.