jenkins: Separate out gke/openshift into 2 separate continuous-upgrade jobs
Also updates the cleanup logic so we can avoid running namespace cleanup
when using testRunner.groovy via continuous-upgrade jobs and updates the
deploy logic to stop uninstalling before install, and to pass the
current time as an annotation to every pod template, ensuring pods are
re-created every deploy.

Finally, the deploy script now correctly attempts to wait for components
to be recreated after the install step.