test/e2e: Move RequireReportDataSourcesForQueryHaveData out of it's own sub-test
Moving this into a subtest prevents it from causing the parent subtest
from bailing out early when collection doesn't have the metrics we want.
So move this out of it's own subTest into the root subTest to prevent
other tests from running when imports fail or zero metric data imported.