Remove generic deploy platform
The operator-metering chart was supposed to be the generic chart for
running metering on non-openshift envs, however it's no longer being
tested, and the only major difference are the default configuration
values. Continuing to maintain it like this takes effort and isn't free
so I'm opting to remove it.

Instead of trying to maintain the other chart, we can just focus
on making openshift-metering work well in openshift, and we can support
non-openshift envs by providing configuration examples that can override
openshift specific options.