pkg/operator: Handle query dependency validation errors more gracefully
Check the type of dependency validation error that occurs and handle it
based on the type of error:

- If it's an uninitialized dependency, stop processing and do not
requeue because dependencies will queue their dependents when they
become initialized.
- If it's an invalid dependency then skip processing, log an error and
do not requeue because it will not succeed anymore unless updated or the
dependency is changed. In both those cases the dependent query will get
queued via the watch, or by dependency after initialization.
- Anything else just bubble up the error to the sync handler and cause
the resource to be dequeued. This happens mostly when unable to get a
dependency.