test/e2e: Update e2e tests to rely on Reports waiting for dependencies.
Instead of waiting on ReportDataSources and ReportGenerationQueries to
become ready reports are created, create the reports and rely on them to
only run once their dependencies are initialized and have data.

Additionally, run the Reports in parallel as it speeds things up now
that the reporting-operator is running regular data imports, it takes
longer to run the tests serially.