images: Write resourceVersion used to Metering CR status
Also updates Dockerfile to use a newer kubectl that is able to patch CRDs.
Currently only kubectl 1.5.2 is available in centos/rhel repos, so we
copy oc from the origin-cli image, and symlink it to kubectl (which is
what the CLI image also does).

The idea is this can be used to determine if a Metering operator has
reacted to an update in the spec, which is useful in CI, and other
situations.