*: Replace metering-helm-operator with metering-ansible-operator
This replaces metering-helm-operator with metering-ansible-operator. The
existing metering-helm-operator Dockerfiles have been updated to just be
`FROM scratch` to avoid breaking CI which still builds these.  We can
remove the Dockerfiles for metering-helm-operator once we update CI to
stop building them.o

The Metering CRD is now MeteringConfig to make it clearer what it's
purpose is.