charts: Fix missing newline in presto configs
The chart config file doesn't always have a trailing newline so
appending to the presto config files in the entrypoint was appending to
the last line, rather than adding a new configuration option on it's own
line.

This fixes this issue by always appending a newline to the config files
we modify before appending any configuration options.