*: Reorganize hdfs to part of a hadoop sub-component for better alignment with configuration usage
We don't use all of hadoop, only HDFS, and additionally, we do use the
hadoop core-site.xml configuration even when not using hdfs. This
reorganizes configuration so that HDFS is more of a sub-section under
hadoop, allowing better re-use of some hadoop configuration in other
components, while not having to deal with HDFS configuration.