charts/openshift-metering: Update openshift-reporting MeteringConfig schema
Updates spec.openshift-reporting.spec.defaultReportDataSources to use a
list instead of a map, and adds an enabled option.

Also uses a helper template to create each ReportDataSource to make it
easier to re-use the ReportDataSource templating logic for when we need
to pick different ReportDataSources based on the version of Kubernetes.