charts,roles/meterinconfig: Fix not creating S3 bucket when spec.storage.hive.s3.createBucket is false
The default() filter isn't correct to use here because it sets the value
to the default when false, or empty when the second argument is true.
This causes the default to be used when the user specifies false.

Fixes that by using _storage_overrides and merging that with the user
configuration to determine if createBucket is true instead. Also updates
other s3 storage fields to use this pattern instead of the default()
filter.