Dockerfile*,images: Fix metering-ansible-operator running in Openshift 3.x
The upstream ansible-operator Dockerfile removed the entrypoint script
which handled adding the container's UID to /etc/passwd as required by
ansible to operate correctly. Modern cri-o releases properly handle this
for us, but we need to continue working with older versions of Openshift
and thus cri-o.

Adds an entrypoint script to our image which handles adding the dynamic
UID to /etc/passwd on startup.