images: Add a _ocp_disabled_overrides dictionary to control non-OKD distributions.
This allows us to handle vanilla Kubernetes by changing the default values for non-OCP users and disabling TLS configurations. In order to use this configuration, you need to either set the environment variable, true, to true, or spec.disableOCPFeatures.enabled in your MeteringConfig to true. By default, the metering operator still uses the OCP values. Also, in this commit changelog is a bunch of whitespace / consistency fixes.