roles/meteringconfig: Use image.override helm value for image overrides
Instead of setting repository/tag separately during image overrides from
the environment variable values, just override the entire image value.

This means we can stop using `split` which has problems when an image
has a port in the URL. When there is a port in the image registry value,
it causes the image to have 2 `:` in the string, causing the split to
split on the port, rather than the image tag in the string, which causes
the entire image tag value to be omitted with the current logic.

This fixes an issue when testing with the internal registry for image
values which is how we test staging OLM package bundles with images from
brew.