hack: Fix mirroring scripts now that internal image registry is different
Images from brew are now stored in quay and accessed through an internal
proxy. As a result, the images are published to a different image
org than before, and the image names no longer map directly between source and final name.
So we need to extract the image tag in get-metering-package-images.sh,
and explicitly set the source image name to the registry proxy image's
name, and set the output image name to what we normally get.