test/e2e: Refactor the e2e package to use the deploy metering package.
With new support for deploying metering using a Golang package (and CLI tool), we can transistion towards removing the hack scripts that deploy metering, and reduce the amount of wrapper scripts used to provision the testing environment. This commit uses the deployframework package to deploy metering, then run the report tests.

Note: we needed to encapsulate the testReportingProducesData test helper function with t.Run in order to respect the defer closure call, and avoid removing the namespace while parallel tests paused/ran/continued.

For more information on the behavior described above, see:
[1] https://github.com/golang/go/issues/22993
[2] https://github.com/golang/go/issues/31651