test/reportingframework: Stat the report output directory.
As we move the process of deploying metering away from the bash hack scripts, and into the testing code, we need to ensure that this directory exists before running any reporting tests. Before, we created the report output directory in the code via `os.MkdirAll(reportOutputDir, 0777)`, getting the value of `reportOutputDir` from an environment variable, and validating that environment variable was set/non-empty.