test/e2e: Fail test install helper function when NewDeployerCtx returns an error.
If there's an error or the DeployerCtx object is nil when running the NewDeployerCtx constructor, we need to exit the test in order to avoid a panic call when the Setup() method is run (as the DeployerCtx object will be nil.)

In the current architecture, we can't return an error from the testInstall helper function, as that function is potentially run in a goroutine (via t.Run(...)), and we need to run the defer closure. If the DeployerCtx object return is nil, then we don't care about logging resource output or uninstalling metering, but we do care about that once the Setup method is run.