test: Update the e2e and deployframework packages to support running the e2e suite locally.
This commit addresses the need to run the e2e suite locally by deploying both the metering and reporting operator locally. In order to this, the deployframework package was extended to support another type that handles the current local context and performs cleanup on any resources provisioned (process IDs, tmp directories, etc.) while running the tests locally.

The test/e2e/main_test.go was also updated to pass a set of extra environment variables that are used when running the reporting-operator locally (but can be lightly refactored to pass to the metering-operator as well). This was needed as the reporting tests have differing behavior in how metrics are imported/managed.