pkg: Support querying the metadata from an existing Presto table when unmanaged.
Currently, spec.columns is a required value in the PrestoTable CRD which is not always ideal. In the case of unmanaged tables, you would still need to specify a spec.columns configuration despite that table schema already existing in Presto. This would update the way that the reporting-operator handles unmanaged tables, namely natively querying the Presto go-client for the spec.tableName table_name, which eliminates the need to specify a column definition.