charts, images: Enable support for auto-detecting single-stack IPV4 or IPV6 networking configurations.
Charts/: This would add a top-level variable that enables IPV6 networking to values.yaml.

Images/: Update the MeteringConfig Ansible role to allow support for inspecting the Networking config.openshift.io object, iterating over the serviceNetwork array and checking if there's any IPV4 addresses. In the case where there's a mixture of IPV4 and IPV6 addresses, treat that configuration as a single-stack IPV4 Openshift cluster. Else, treat it as a single-stack IPV6 Openshift cluster.