charts,images: Add the AWS region to the storage overrides dictionary.
This is necessary as storage.hive is considered an interface in the values.yaml file, so attempting to index the AWS region from storage.hive.s3.region is not possible and will fail the "helm template ..." stage of the "make verify" test.

In order to get around this, we need to add the "region" key to the Presto/Hadoop S3 storage overrides dictionary and update the values.yaml to include this key. Then, we can reference this value in the corresponding helm charts.