charts: Update the Hive and reporting-operator helm charts to work with single-stack ipv4 and ipv6.
This is a follow-up to PR #1077 which introduced support for using S3 as the storage backend in a IPV6 cluster. The changes in this PR update the Hive ghostunnel configuration needed to support both a single-stack configuration, dependent on the value of the top-level .useIPV6Networking.

In the case of the pkg/reporting-operator/cmd.go flags, we typically override the default values through the corresponding environment variable, which is reflected in the reporting-operator deployment changes. It's important to note that this means we need to be more careful when interacting with the CLI explicitly (i.e. in the case of running the reporting-operator locally, or in the e2e suite.)