pkg: Support creating unmanaged PrestoTables from a ReportDataSource.
This would add the ability to create a ReportDataSource that references an existing table in Presto. Before, you would need to first create an unmanaged PrestoTable that referenced an existing table in Presto, then create a ReportDataSource that referenced the created PrestoTable. With initial support for the presto-prometheus connector being integrated, this change would be needed as Presto would read direcly from Prometheus and store relevant data into tables which ReportDataSources can then reference.