Dockerfile: Add the python botocore package to the metering-ansible-operator images.
This is needed as the s3_bucket Ansible module does not support dualstack endpoints, which is problematic is metering is run in an ipv6 single-stack configuration. In order to use the aws_s3 module which exposes a 'dualstack' option, but has a dependency on the botocore python package. This commit would install that package so we can utilize this module to support both single-stack ipv4 and single-stack ipv6 configurations.