images: Replace the s3_bucket Ansible module with the aws_s3 module.
This commit would replace the s3_bucket module with the aws_s3 module with the idea that we can support for single-stack ipv4 and ipv6 with the aws_s3 module. In the current scope, the s3_bucket does not have the capability of using the dualstack AWS API endpoint, which is results in an error in the role execution and no s3 bucket being successfully created.