images: Move the configure networking task file to execute earlier in the role execution.
The configuring networking task file initializes the value of the _ocp_enabled_use_ipv6_networking by analyzing the cluster networking API object and is going to be needed to be used throughout the role. In the case of configuring storage, we need to determine whether or not to use this 'dualstack' option which is contingent on the value of the _ocp_enabled_use_ipv6_networking value.