pkg: Support creating unamanged PrestoTables from a ReportDataSource.
This would add the ability to create a ReportDataSource that references an existing table in Presto. Before, you would need to first create an unmanaged PrestoTable that referenced an existing table in Presto, then create a ReportDataSource using the spec.prestoTable sub-type. With initial support for the presto-prometheus connector being integrated, this change would be needed as Presto would read directly from Prometheus and store relevant data into tables which the ReportDataSources can then reference.