Dockerfile,images: Remove the metering-ansible-operator entrypoint script.
This was the recommended way to support arbitrary user IDs in pre-4.2 releases as the container UID was generated dynamically and may not have the corresponding /etc/passwd entry.

In 4.2+ releases, CRI-O automatically updates this /etc/passwd entry with the current container UID so this is now obsolete and insecure.