Dockerfile: Update the metering-ansible-operator to use the Ansible 2.8 and newer versions.
Previously, we needed to always use the Ansible 2.8 version in the metering-ansible-operator origin Dockerfile as there was a regression in 2.9 with regards to the 'ansible_failed_result' and 'ansible_failed_task' variables being out of scope in a block/rescue task. It would appear this is no longer necessary and we can use the Ansible 2.9 version.