charts,images: Add support for auto-detecting when a cluster-wide Proxy has been configured.
Metering currently does not respect any cluster-wide Proxy configurations, so this would be the first step towards implementing support.

On the Metering side-of-things, when this Proxy cluster object is not nil, we need to set a top-level boolean variable that can be used
at the Helm chart level to determine whether or not to use a certain proxy configuration or not. In addition to that, we need to update
our cluster role/rolebinding permissions which allows us to directly read from this cluster-scope Proxy object, pulling out information
we need and overriding the default proxy configuration values that Metering reads from.