images: Add a hive metastore database overrides dictionary.
By default, hive metastore creates a local, embedded derby database PVC that gets mounted to the Hive metastore Pod. In the case where a user wants to use a database like MySQL or PostgreSQL, they would need to specify this information in the MeteringConfig custom resource, and hive metastore would in turn use this dedicated SQL database for storing metastore data (and not create the derby PVC). In the current state of things, this PVC is still being created, despite a database configuration being provided in the user's MeteringConfig custom resource.