cmd,images: Connect to the thanos-querier API instead of the cluster-wide Prometheus instance.
In 4.3, the monitoring team updated its architecture to introduce this Thanos Querier component, which provides a way for components like Metering to connect to a single, multi-tenant interface.
By default and when run on Openshift, Metering connects to the cluster-wide, prometheus-k8s service to retrieve the metrics scraped from the cluster. These changes update the metering-ansible-operator Ansible role to connect to the thanos-querier service API instead of the prometheus-k8s service.

In the case of running Metering on a vanilla Kubernetes cluster, nothing has changed - you're still required to configure/provide the Prometheus URL in the `MeteringConfig` custom resource.