pkg/deploy: Fix the hive-metastore PVC label selector.
When uninstalling the metering associated PVCs, we list all the PVCs in the deploy namespace that match the app=hdfs or app=hive. This is problematic as the hive-metastore PVC chart does not match the rest of the Hive components labeling structure (app=hive,hive=<metastore,server>) so we are label selecting something that won't ever exist.